calendar.php js code integrated fully with jQuery and decomposed
